Allow single quotes in import statements
Currently, an import error is thrown when attempting to use single
quotes in the filename string in an import statement. Even though
the import statement itself isn't strictly legal javascript, it may
be helpful to permit both single and double quoted strings to better
fit the end-users preferred quoting style.

Introduce minor change to the regex pattern in emit-imports.js to
test for both single and double-quoted strings. Use back-reference
to ensure that the same delimiter is used on both ends of the string.
Thus, "path" and 'path' are valid but "path' and 'path" are not.